requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
Pillow>=9.0.0
easyocr>=1.7.0
